+++ backend/src/app.module.ts
@@ -21,12 +21,8 @@ import { APP_GUARD } from "@nestjs/core";
 import { RolesGuard } from "./auth/guards/roles.guard";
 import { LikeModel } from "./post/models/like.model";
 import { DislikeModel } from "./post/models/dislike.model";
-import { MessageGateway } from "./message/message.gateway";
 import { MessageModule } from "./message/message.module";
-import { UserConversationModel } from "./message/models/user-conversation.model";
-import { ConversationModel } from "./message/models/conversation.model";
-import { MessageModel } from "./message/models/message.model";
-import { ActiveConversationModel } from "./message/models/active-conversation";
+import { MessageGateway } from "./message/message.gateway";
 
 @Module({
   providers: [
@@ -60,10 +56,6 @@ import { ActiveConversationModel } from "./message/models/active-conversation";
         PhotoModel,
         LikeModel,
         DislikeModel,
-        UserConversationModel,
-        ConversationModel,
-        MessageModel,
-        ActiveConversationModel,
       ],
       autoLoadModels: true,
       synchronize: true,