+++ backend/src/post/models/post.model.ts
@@ -1,17 +1,8 @@
-import {
-  BelongsTo,
-  Column,
-  DataType,
-  ForeignKey,
-  HasMany,
-  Model,
-  Table,
-} from "sequelize-typescript";
+import { Column, DataType, ForeignKey, HasMany, Model, Table } from "sequelize-typescript";
 import { ApiProperty } from "@nestjs/swagger";
 import { UserModel } from "../../user/models/user.model";
 import { DislikeModel } from "./dislike.model";
 import { LikeModel } from "./like.model";
-import { options } from "tsconfig-paths/lib/options";
 
 interface PostCreationAttrs {
   title: string;