+++ backend/src/user/models/follow.model.ts
@@ -21,10 +21,10 @@ export class FollowModel extends Model<FollowModel, FollowAttrCreation> {
   @ApiProperty({ example: "19", description: "User's identifier" })
   @ForeignKey(() => UserModel)
   @Column({ type: DataType.INTEGER })
-  user_id: number;
+  userId: number;
 
   @ApiProperty({ example: "32", description: "Follower's identifier" })
   @ForeignKey(() => UserModel)
   @Column({ type: DataType.INTEGER })
-  follower_id: number;
+  followerId: number;
 }