+++ frontend/src/types/profile/follow.dto.ts
@@ -1,5 +1,5 @@
 export interface FollowDto {
-    user_id: number;
-    follow_id: number;
+    userId: number;
+    followId: number;
 
 }